requirements.txt
requests
minio
netifaces
kafka-python